+++ ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: seohyeki <seohyeki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/10 15:22:08 by seohyeki          #+#    #+#             */
-/*   Updated: 2023/10/31 14:33:05 by seohyeki         ###   ########.fr       */
+/*   Updated: 2023/11/07 16:33:19 by seohyeki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 static char	*ft_strcat(char *dest, const char *src)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
@@ -33,7 +33,7 @@ static char	*ft_strcat(char *dest, const char *src)
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
-	int		len;
+	size_t	len;
 	char	*new_str;
 
 	len = ft_strlen(s1) + ft_strlen(s2) + 1;
