+++ ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: seohyeki <seohyeki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/05 19:38:06 by seohyeki          #+#    #+#             */
-/*   Updated: 2023/11/06 14:26:42 by seohyeki         ###   ########.fr       */
+/*   Updated: 2023/11/07 16:36:05 by seohyeki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (*s != '\0')
